Split up app version utility functions
