+++ src/utils/app-version/get-develop-build.ts
@@ -1,0 +1,5 @@
+import { DevBuild, getDevBuildBranch } from '.';
+
+export default function getDevelopBuild(): Promise<DevBuild> {
+  return getDevBuildBranch('develop');
+}
