+++ src/utils/app-version/index.ts
@@ -1,0 +1,6 @@
+export * from './types';
+export { default as getBetaBuild } from './get-beta-build';
+export { default as getDevBranchBuilds } from './get-dev-branch-builds';
+export { default as getDevBuildBranch } from './get-dev-build-branch';
+export { default as getDevelopBuild } from './get-develop-build';
+export { default as getPrBuilds } from './get-pr-builds';
